Move shifting code out of expr and into somewhere more accessible
